修复排行榜,个人排名等级排名错误的BUG, 应该用UInt64而不是Int

+++ Core/GObject/Leaderboard.h
@@ -129,7 +129,7 @@ private:
 	std::map<UInt32, UInt16> _clanRankCountry[2];
     std::vector<UInt64> _levelRankWorld10;  //世界等级前十名
     std::multimap<int, Player*, std::greater<int> > _battleRankWorld;
-    std::multimap<UInt64, Player*, std::greater<int> > _expRankWorld;
+    std::multimap<UInt64, Player*, std::greater<UInt64> > _expRankWorld;
     FastMutex _tmutex;
     std::vector<LeaderboardTowndown> _towndown;
     FastMutex _cmutex;
